Add files via upload
To Do List To Do List is a project based on HTML, CSS and Javascript which helps people writing down a list of tasks that they want to get done, or that need to get done. All you have to do is open the .html file and start typing. Once you're finished typing the task, you just press Enter to add it to the list and you're good to go with the next one. Click the 'X' button to remove the task.